Changing the timer to tick each second
Don't know if this was proposed a while ago or not, but for my use case this fixes the issue I had. I can try making it so there's a new option on the menu for ticking each second instead of after the 10s mark, but for now this was the "easiest" solution.